+++ platforms/linux/riscv64-clang.toolchain.cmake
@@ -22,10 +22,10 @@ set(CMAKE_CXX_FLAGS "-march=rv64gcv --gcc-toolchain=${RISCV_GCC_INSTALL_ROOT} -w
 
 set(CMAKE_C_FLAGS_RELEASE "${CMAKE_C_FLAGS_RELEASE}  -O2")
 set(CMAKE_CXX_FLAGS_RELEASE "${CMAKE_CXX_FLAGS_RELEASE}  -O2")
-OPTION(NEWRVV "use new rvv ui" ON) # Enabled by default
-IF(NEWRVV)
-    ADD_DEFINITIONS(-DCV_RVV_NEW)
-ENDIF(NEWRVV)
+OPTION(RISCV_RVV_SCALABLE "Use scalable RVV API on RISC-V" ON) # Enabled by default
+IF(RISCV_RVV_SCALABLE)
+    ADD_DEFINITIONS(-DCV_RVV_SCALABLE)
+ENDIF()
 
 set(CMAKE_FIND_ROOT_PATH ${CMAKE_SYSROOT})
 set(CMAKE_FIND_ROOT_PATH_MODE_PROGRAM NEVER)
